Extract fakeGCR to use both Copy and Delete testing.
